Introduce BigScreen styled FullScreenOverlay & update Favourites Manager style
This MR introduces a new FullScreenOverlay component, based on the NanoShell.FullScreenOverlay. It is a styled full screen shell overlay component with a header, close button and animation. It is intended to be shared across similar full screen overlays within the shell.

The first usage of this overlay is in the Favourites Manager, which has been slightly modified visually and now includes add and remove icons.

Favourites Manager Before:
![Screenshot_20250324_233134](/uploads/f28b1742950c3f9569d810edba6f92f1/Screenshot_20250324_233134.png){width=1358 height=852}

Favourites Manager After:
![Screenshot_20250325_233454](/uploads/80fffbf0e4e9f37d61c70c141e64018d/Screenshot_20250325_233454.png){width=1358 height=852}