kcms/bigscreen-settings: Port to new components, and fixes
This MR overhauls the bigscreen-settings KCM. It appears that many of
the settings were removed in a previous MR:
https://invent.kde.org/plasma/plasma-bigscreen/-/merge_requests/102,
this restores them ported to the new components.

This MR also changes the name of the KCM to "System" to better reflect
that it deals with several system options as well as appearance ones.

Some bugs were also fixed:
- Fixed shortcut reset not being implemented
- Fixed timezone setting in the time sidebar
- Fixed the listed formfactors (should only be mediacenter)
- KeyNavigation.left is now set on KCM pages by ConfigWindow

![Screencast_20250622_080917](/uploads/899426cccc3cdcfe810af0c0b8fea87c/Screencast_20250622_080917.webm)