settingsapp: Go back to checking FormFactors instead of KCM id
**Closes: #47**

Previously, due to the lack of `mediacenter` in KCM `FormFactors`, the way to check if a KCM belonged in Settings was by checking if the ID had `mediacenter` in it. This method is obsolete now, because the `mediacenter` form factor has been adopted across KDE.

The problem right now is that the `mediacenter` form factor applies to way too many KCMs, even Plasma Mobile ones, so it has to be removed from them.

**Code changes:**
- [x] In `kcmlistmodel.cpp`, check if form factors contains `mediacenter` instead of the plugin ID